refactoring table layouts dir structure
